mejora
implemente el modelo por separado
