+++ actividad3/erick/garcia/process/N8_PromedioSalario.java
@@ -3,6 +3,8 @@ package erick.garcia.process;
 import java.util.Arrays;
 import java.util.List;
 
+import erick.garcia.models.Empleado;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Assertions;
